requirements.txt
python-dotenv==1.1.0
telethon==1.40.0